Modularize repos.rs into submodules for maintainability
Split src/db/repos.rs (513 lines) into domain-focused submodules:
- watched.rs: WatchedRepo and watched repos functions
- links.rs: RepoLink and forge link functions
- worktree.rs: worktree issue associations
- tests.rs: all tests for repos module
